// ng-ncc-app/src/app/interfaces/contact-details.interface.ts
export interface ContactDetails {
    contactId: string;
    emailAddress: string;
    uprn: string;
    addressLine1: string;
    addressLine2: string;
    addressLine3: string;
    firstName: string;
    lastName: string;
    larn: string | null;
    address1AddressId: string;
    address2AddressId: string;
    address3AddressId: string;
    telephone1: string | null;
    telephone2: string | null;
    telephone3: string | null;
    cautionaryAlert: boolean;
    propertyCautionaryAlert: boolean;
    houseRef: string | null;
    title: string | null;
    fullAddressDisplay: string;
    fullAddressSearch: string;
    postCode: string;
    dateOfBirth: Date;
    hackneyHomesId: string;
    houseHoldId: string;
    memberId: string;
    personno: number;
    accountId: string;
}